Encapsulate the Star-Planet Code in terrasol.py
I moved most of the bokeh plot creation and updating inside of a class
in terrasol.py.  This will allow for more of a plug and play nature
with modules that will show up on the web page.
